+++ code/messaging/src/main/java/com/adobe/marketing/mobile/aepcomposeui/uimodels/AepColor.kt
@@ -9,7 +9,7 @@
   governing permissions and limitations under the License.
 */
 
-package com.adobe.marketing.mobile.aepuitemplates.uimodels
+package com.adobe.marketing.mobile.aepcomposeui.uimodels
 
 /**
  * Represents the colors used for UI elements in both light and dark themes.
